Add files via upload
An interactive game that the player can challenge the computer with 3 levels: easy, mid and master, also available to choose who plays first and end the game at any time. 
It would automatically compute who wins the game this time and restart the game if the player wants to play again. Once the player ends it, a statistic of how many games the player has won/ lose/ draw would show up on the screen. This could be used in further data analysis. 

#Modules used:
Numpy (3rd-party):  Speed up and calculate complex algorithm.
Random (Built-in):  Increase unpredictable moves in easy level and some of mid & master level.